fix bug

+++ baby-back-backend/springcloud-admin/springcloud-admin-api/src/main/java/com/babyback/api/admin/api/config/SctUser.java
@@ -16,9 +16,9 @@ public class SctUser extends User {
      * 用户ID
      */
     @Getter
-    private Long id;
+    private Integer id;
 
-    public SctUser(Long id, String username, String password, boolean enabled, boolean accountNonExpired, boolean credentialsNonExpired, boolean accountNonLocked, Collection<? extends GrantedAuthority> authorities) {
+    public SctUser(Integer id, String username, String password, boolean enabled, boolean accountNonExpired, boolean credentialsNonExpired, boolean accountNonLocked, Collection<? extends GrantedAuthority> authorities) {
         super(username, password, enabled, accountNonExpired, credentialsNonExpired, accountNonLocked, authorities);
         this.id = id;
     }
